requirements.txt
fastapi
tortoise-orm[asyncpg]
pydantic
pydantic-settings
aerich
ipython
rich
